personalized the database

+++ database.py
@@ -49,89 +49,89 @@ class Database:
     def __init__(self, dbfile):
         self.dbfile = dbfile
 
-    def add_participant(self, participant):
+    def add_participant(self, user_id, participant):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "INSERT INTO PARTICIPANT (PARTICIPANT, EMAIL) VALUES (?, ?)"
-            cursor.execute(query, (participant.participant, participant.email))
+            query = "INSERT INTO PARTICIPANT (user_id, PARTICIPANT, EMAIL) VALUES (?, ?, ?)"
+            cursor.execute(query, (user_id, participant.participant, participant.email))
             connection.commit()
             participant_key = cursor.lastrowid
         return participant_key
 
 
-    def update_participant(self, participant_key, participant):
+    def update_participant(self, user_id, participant_key, participant):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "UPDATE PARTICIPANT SET PARTICIPANT = ?, EMAIL = ? WHERE (ID = ?)"
-            cursor.execute(query, (participant.participant, participant.email, participant_key))
+            query = "UPDATE PARTICIPANT SET PARTICIPANT = ?, EMAIL = ? WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (participant.participant, participant.email, participant_key, user_id))
             connection.commit()
     
-    def delete_participant(self, participant_key):
+    def delete_participant(self, user_id, participant_key):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "DELETE FROM PARTICIPANT WHERE (ID = ?)"
-            cursor.execute(query, (participant_key,))
+            query = "DELETE FROM PARTICIPANT WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (participant_key, user_id))
             connection.commit()
     
-    def get_participant(self, participant_key):
+    def get_participant(self, user_id, participant_key):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT PARTICIPANT, EMAIL FROM PARTICIPANT WHERE (ID = ?)"
-            cursor.execute(query, (participant_key,))
+            query = "SELECT PARTICIPANT, EMAIL FROM PARTICIPANT WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (participant_key, user_id))
             participant, email = cursor.fetchone()
         participant_ = Participant(participant, email=email)
         return participant_
     
-    def get_participants(self):
+    def get_participants(self, user_id):
         participants = []
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT ID, PARTICIPANT, EMAIL FROM PARTICIPANT ORDER BY ID"
-            cursor.execute(query)
+            query = "SELECT ID, PARTICIPANT, EMAIL FROM PARTICIPANT WHERE (user_id = ?) ORDER BY ID"
+            cursor.execute(query, (user_id,))
             for participant_key, participant, email in cursor:
                 participants.append((participant_key, Participant(participant, email)))
         return participants
 
     ############## COUPLES #####################
 
-    def add_couple(self, couple):
+    def add_couple(self, user_id, couple):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "INSERT INTO COUPLE (PARTNERONE, PARTNERTWO) VALUES (?, ?)"
-            cursor.execute(query, (couple.partner_one, couple.partner_two))
+            query = "INSERT INTO COUPLE (user_id, PARTNERONE, PARTNERTWO) VALUES (?, ?, ?)"
+            cursor.execute(query, (user_id, couple.partner_one, couple.partner_two))
             connection.commit()
             couple_key = cursor.lastrowid
         return couple_key
     
-    def update_couple(self, couple_key, couple):
+    def update_couple(self, user_id, couple_key, couple):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "UPDATE COUPLE SET PARTNERONE = ?, PARTNERTWO = ? WHERE (ID = ?)"
-            cursor.execute(query, (couple.partner_one, couple.partner_two, couple_key))
+            query = "UPDATE COUPLE SET PARTNERONE = ?, PARTNERTWO = ? WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (couple.partner_one, couple.partner_two, couple_key, user_id))
             connection.commit()
 
-    def delete_couple(self, couple_key):
+    def delete_couple(self, user_id, couple_key):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "DELETE FROM COUPLE WHERE (ID = ?)"
-            cursor.execute(query, (couple_key,))
+            query = "DELETE FROM COUPLE WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (couple_key, user_id))
             connection.commit()
     
-    def get_couple(self, couple_key):
+    def get_couple(self, user_id, couple_key):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT PARTNERONE, PARTNERTWO FROM COUPLE WHERE (ID = ?)"
-            cursor.execute(query, (couple_key,))
+            query = "SELECT PARTNERONE, PARTNERTWO FROM COUPLE WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (couple_key, user_id))
             partner_one, partner_two = cursor.fetchone()
         couple_ = Couple(partner_one, partner_two=partner_two)
         return couple_
     
-    def get_couples(self):
+    def get_couples(self, user_id):
         couples = []
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT ID, PARTNERONE, PARTNERTWO FROM COUPLE ORDER BY ID"
-            cursor.execute(query)
+            query = "SELECT ID, PARTNERONE, PARTNERTWO FROM COUPLE WHERE (user_id = ?) ORDER BY ID"
+            cursor.execute(query, (user_id,))
             for couple_key, partner_one, partner_two in cursor:
                 couples.append((couple_key, Couple(partner_one, partner_two)))
         return couples
@@ -140,60 +140,60 @@ class Database:
 
     ############## ASSIGNMENTS #####################
 
-    def add_assignment(self, assignment):
+    def add_assignment(self, user_id, assignment):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "INSERT INTO ASSIGNMENT (NAMEONE, NAMETWO) VALUES (?, ?)"
-            cursor.execute(query, (assignment.name1, assignment.name2))
+            query = "INSERT INTO ASSIGNMENT (user_id, NAMEONE, NAMETWO) VALUES (?, ?, ?)"
+            cursor.execute(query, (user_id, assignment.name1, assignment.name2))
             connection.commit()
             assignment_key = cursor.lastrowid
         return assignment_key
     
-    def update_assignment(self, assignment_key, assignment):
+    def update_assignment(self, user_id, assignment_key, assignment):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "UPDATE ASSIGNMENT SET NAMEONE = ?, NAMETWO = ? WHERE (ID = ?)"
-            cursor.execute(query, (assignment.name1, assignment.name2, assignment_key))
+            query = "UPDATE ASSIGNMENT SET NAMEONE = ?, NAMETWO = ? WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (assignment.name1, assignment.name2, assignment_key, user_id))
             connection.commit()
     
-    def delete_assignment(self, assignment_key):
+    def delete_assignment(self, user_id, assignment_key):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "DELETE FROM ASSIGNMENT WHERE (ID = ?)"
-            cursor.execute(query, (assignment_key,))
+            query = "DELETE FROM ASSIGNMENT WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (assignment_key, user_id))
             connection.commit()
 
-    def get_assignment(self, assignment_key):
+    def get_assignment(self, user_id, assignment_key):
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT NAMEONE, NAMETWO FROM ASSIGNMENT WHERE (ID = ?)"
-            cursor.execute(query, (assignment_key,))
+            query = "SELECT NAMEONE, NAMETWO FROM ASSIGNMENT WHERE (ID = ?) AND (user_id = ?)"
+            cursor.execute(query, (assignment_key, user_id))
             name1, name2 = cursor.fetchone()
         assignment_ = Assignment(name1, name2=name2)
         return assignment_
 
-    def get_assignments(self):
+    def get_assignments(self, user_id):
         assignments = [] 
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT ID, NAMEONE, NAMETWO FROM ASSIGNMENT ORDER BY ID"
-            cursor.execute(query)
+            query = "SELECT ID, NAMEONE, NAMETWO FROM ASSIGNMENT WHERE (user_id = ?) ORDER BY ID"
+            cursor.execute(query, (user_id,))
             for assignment_key, name1, name2 in cursor:
                 assignments.append((assignment_key, Assignment(name1, name2)))
         return assignments
 
-    def get_latest_assignments(self):
+    def get_latest_assignments(self, user_id):
         assignments = [] 
         with sqlite3.connect(self.dbfile) as connection:
             cursor = connection.cursor()
-            query = "SELECT ID, PARTICIPANT, EMAIL FROM PARTICIPANT ORDER BY ID"
-            cursor.execute(query)
+            query = "SELECT ID, PARTICIPANT, EMAIL FROM PARTICIPANT WHERE (user_id = ?) ORDER BY ID"
+            cursor.execute(query, (user_id,))
             number_of_latest_assignments = 0
             # use the number of participants to get the number of assignments
             for participant_key, participant, email in cursor:
                 number_of_latest_assignments += 1
-            query = f"SELECT ID, NAMEONE, NAMETWO FROM ASSIGNMENT ORDER BY ID DESC LIMIT {number_of_latest_assignments}"
-            cursor.execute(query)
+            query = f"SELECT ID, NAMEONE, NAMETWO FROM ASSIGNMENT WHERE (user_id = ?) ORDER BY ID DESC LIMIT {number_of_latest_assignments}"
+            cursor.execute(query, (user_id,))
             for assignment_key, name1, name2 in cursor:
                 print(assignment_key, name1, name2)
                 assignments.append((assignment_key, Assignment(name1, name2)))
